Various changes to prevent narrowing conversions: components/
Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2957543
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892623}

+++ components/flags_ui/flags_state.cc
@@ -623,7 +623,7 @@ void FlagsState::GetFlagFeatureEntries(
 }
 
 // static
-int FlagsState::GetCurrentPlatform() {
+unsigned short FlagsState::GetCurrentPlatform() {
 #if defined(OS_IOS)
   return kOsIos;
 #elif defined(OS_MAC)
